requirements.txt
chromadb==0.4.8
faiss-cpu==1.7.4
fastapi==0.96.0
langchain==0.0.251
mlflow==2.7.0
nemoguardrails==0.5.0
openai==0.27.8
pdfminer-six==20221105
PyMuPDF==1.23.3
pysqlite-binary
python-dotenv==1.0.0
streamlit==1.25.0
streamlit-chat==0.1.1
tiktoken==0.4.0
trubrics==1.4.3
watchdog==3.0.0